feat: Add fetchFeaturedBooks method to HomeRepoImpl
The fetchFeaturedBooks method in HomeRepoImpl is implemented to fetch a list of featured books related to the subject "islam" using the ApiService. It handles successful responses by converting the data into a list of BookModel objects and returns it. If there is an error, it handles DioException by converting it into a ServerFailure and returns it.